requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
supabase
python-dotenv
assemblyai
